requirements.txt
httpx>=0.27.0
wandb>=0.16.0
weave>=0.51.0
Pillow>=10.0.0
python-dotenv>=1.0.0
